storage/requirements.txt
requests
pika
azure-storage-blob